requirements.txt
requests>=2.31.0
types-requests>=2.31.0
python-dotenv>=0.21.0
fitparse>=1.2.0
aiohttp>=3.8.0